+++ tests/test_image_conversion_backend.py
@@ -4,18 +4,16 @@ import pytest
 
 import camelot.backends.image_conversion
 from camelot.backends import ImageConversionBackend
-from camelot.backends.poppler_backend import PopplerBackend
-from camelot.backends.ghostscript_backend import  GhostscriptBackend
 
 
 class PopplerBackendError(object):
     def convert(self, pdf_path, png_path):
-        raise ValueError('conversion failed')
+        raise ValueError("Image conversion failed")
 
 
 class GhostscriptBackendError(object):
     def convert(self, pdf_path, png_path):
-        raise ValueError('conversion failed')
+        raise ValueError("Image conversion failed")
 
 
 class GhostscriptBackendNoError(object):
@@ -24,26 +22,39 @@ class GhostscriptBackendNoError(object):
 
 
 def test_poppler_backend_error_when_no_use_fallback(monkeypatch):
-    backends = {"poppler": PopplerBackendError, "ghostscript": GhostscriptBackendNoError}
-    monkeypatch.setattr("camelot.backends.image_conversion.backends", backends, raising=True)
+    BACKENDS = {
+        "poppler": PopplerBackendError,
+        "ghostscript": GhostscriptBackendNoError,
+    }
+    monkeypatch.setattr(
+        "camelot.backends.image_conversion.BACKENDS", BACKENDS, raising=True
+    )
     backend = ImageConversionBackend(use_fallback=False)
 
-    message = "conversion failed with image conversion backend 'poppler'"
+    message = "Image conversion failed with image conversion backend 'poppler'"
     with pytest.raises(ValueError, match=message):
-        backend.convert('foo', 'bar')
+        backend.convert("foo", "bar")
+
 
 def test_ghostscript_backend_when_use_fallback(monkeypatch):
-    backends = {"poppler": PopplerBackendError, "ghostscript": GhostscriptBackendNoError}
-    monkeypatch.setattr("camelot.backends.image_conversion.backends", backends, raising=True)
+    BACKENDS = {
+        "poppler": PopplerBackendError,
+        "ghostscript": GhostscriptBackendNoError,
+    }
+    monkeypatch.setattr(
+        "camelot.backends.image_conversion.BACKENDS", BACKENDS, raising=True
+    )
     backend = ImageConversionBackend()
-    backend.convert('foo', 'bar')
+    backend.convert("foo", "bar")
 
 
 def test_ghostscript_backend_error_when_use_fallback(monkeypatch):
-    backends = {"poppler": PopplerBackendError, "ghostscript": GhostscriptBackendError}
-    monkeypatch.setattr("camelot.backends.image_conversion.backends", backends, raising=True)
+    BACKENDS = {"poppler": PopplerBackendError, "ghostscript": GhostscriptBackendError}
+    monkeypatch.setattr(
+        "camelot.backends.image_conversion.BACKENDS", BACKENDS, raising=True
+    )
     backend = ImageConversionBackend()
 
-    message = "conversion failed with image conversion backend 'ghostscript'"
+    message = "Image conversion failed with image conversion backend 'ghostscript'"
     with pytest.raises(ValueError, match=message):
-        backend.convert('foo', 'bar')
+        backend.convert("foo", "bar")