+++ tests/test_stream.py
@@ -2,6 +2,7 @@
 
 import os
 
+import pytest
 import pandas as pd
 from pandas.testing import assert_frame_equal